requirements.txt
tqdm
pandas 
cassandra-driver 
kafka-python
faker